# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IRIS API Server that supports various light clients",
        "title": "IRIS SERVER API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "host": "host",
    "paths": {
        "/candidates": {
            "get": {
                "description": "the list of candidates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stake"
                ],
                "summary": "candidate list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "per_page",
                        "name": "per_page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "content of data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/document.Candidate"
                            }
                        }
                    }
                }
            }
        },
        "/delegators/{address}/candidates": {
            "get": {
                "description": "list of candidate which delegated by address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stake"
                ],
                "summary": "delegator candidate list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "per_page",
                        "name": "per_page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "order",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "content of data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/document.Candidate"
                            }
                        }
                    }
                }
            }
        },
        "/delegators/{address}/candidates/{pub_key}": {
            "get": {
                "description": "detail of candidate which delegated by address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stake"
                ],
                "summary": "delegator candidate detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "public key of candidate",
                        "name": "pub_key",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "content of data",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/document.Candidate"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "document.Candidate": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "delegators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/document.Delegator"
                    }
                },
                "description": {
                    "type": "object",
                    "$ref": "#/definitions/document.Description"
                },
                "pub_key": {
                    "type": "string"
                },
                "shares": {
                    "type": "integer"
                },
                "update_time": {
                    "type": "string"
                },
                "voting_power": {
                    "type": "number"
                }
            }
        },
        "document.Delegator": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "pub_key": {
                    "type": "string"
                },
                "shares": {
                    "type": "integer"
                },
                "update_time": {
                    "type": "string"
                }
            }
        },
        "document.Description": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "identity": {
                    "type": "string"
                },
                "moniker": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        }
    }
}